requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Config & validation ─────────────────────────────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# ── Database (async PostgreSQL) ─────────────────────────────────────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# ── Groq LLM ────────────────────────────────────────────────────────────────────
groq==0.11.0

# ── RAG / Vector embeddings ─────────────────────────────────────────────────────
pinecone-client==4.1.0
sentence-transformers==3.1.1

# ── HTTP client (RapidAPI + weather) ───────────────────────────────────────────
httpx==0.27.2
